Little fixes and adding  schema.sql and data.sql to initialize the database

+++ src/main/java/com/kata/Booking/API/domain/Hotel.java
@@ -3,4 +3,4 @@ package com.kata.Booking.API.domain;
 import java.util.List;
 
 
-public record  Hotel(Integer id, String name, List<Room> rooms) {}
+public record  Hotel(Long id, String name, List<Room> rooms) {}
